// src/app/generated/model/documentCreateUpdateDTO.ts
/**
 * document-management
 * 1000kit document management
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RelatedPartyRefCreateUpdateDTO } from './relatedPartyRefCreateUpdateDTO';
import { DocumentRelationshipCreateUpdateDTO } from './documentRelationshipCreateUpdateDTO';
import { AttachmentCreateUpdateDTO } from './attachmentCreateUpdateDTO';
import { RelatedObjectRefCreateUpdateDTO } from './relatedObjectRefCreateUpdateDTO';
import { CategoryCreateUpdateDTO } from './categoryCreateUpdateDTO';
import { LifeCycleState } from './lifeCycleState';
import { DocumentCharacteristicCreateUpdateDTO } from './documentCharacteristicCreateUpdateDTO';
import { ChannelCreateUpdateDTO } from './channelCreateUpdateDTO';
import { DocumentSpecificationCreateUpdateDTO } from './documentSpecificationCreateUpdateDTO';

export interface DocumentCreateUpdateDTO {
  id?: string;
  attachments?: Array<AttachmentCreateUpdateDTO>;
  categories?: Array<CategoryCreateUpdateDTO>;
  channel: ChannelCreateUpdateDTO & object;
  characteristics?: Array<DocumentCharacteristicCreateUpdateDTO>;
  description?: string;
  documentRelationships?: Array<DocumentRelationshipCreateUpdateDTO>;
  documentVersion?: string;
  lifeCycleState?: LifeCycleState;
  name?: string;
  relatedObject?: RelatedObjectRefCreateUpdateDTO;
  relatedParties?: Array<RelatedPartyRefCreateUpdateDTO>;
  specification?: DocumentSpecificationCreateUpdateDTO & object;
  tags?: Array<string>;
  typeId: string;
}